#195 Add sourceonly option for python installer

+++ installer/install_python.sh
@@ -5,16 +5,13 @@
 #
 #  Maintainer: id774 <idnanashi@gmail.com>
 #
+#  v1.1 2/21,2009
+#       Add sourceonly option.
 #  v1.0 1/7,2009
 #       Stable.
 ########################################################################
 
-install_python_bzip() {
-    mkdir install_python
-    cd install_python
-    wget http://www.python.org/ftp/python/$1/Python-$1.tar.bz2
-    test -f Python-$1.tar.bz2 || exit 1
-    tar xjvf Python-$1.tar.bz2
+make_and_install() {
     cd Python-$1
     test -n "$2" || ./configure
     test -n "$2" && ./configure --prefix $2
@@ -22,6 +19,15 @@ install_python_bzip() {
     test -n "$2" || sudo make install
     test -n "$2" && make install
     cd ..
+}
+
+get_python() {
+    mkdir install_python
+    cd install_python
+    wget http://www.python.org/ftp/python/$1/Python-$1.tar.bz2
+    test -f Python-$1.tar.bz2 || exit 1
+    tar xjvf Python-$1.tar.bz2
+    test "$2" = "sourceonly" || make_and_install
     test -d /usr/local/src/python || sudo mkdir -p /usr/local/src/python
     sudo cp $OPTIONS Python-$1 /usr/local/src/python
     cd ..
@@ -38,7 +44,7 @@ case $OSTYPE in
 esac
 
 test -n "$1" || exit 1
-install_python_bzip $1 $2
+get_python $1 $2
 
 case $OSTYPE in
   *darwin*)
